Conflicts:
	daemon/src/audio/audiolayer.cpp
	daemon/src/audio/audiortp/audio_rtp_record_handler.cpp
	daemon/src/audio/codecs/alaw.cpp
	daemon/src/audio/codecs/audiocodec.cpp
	daemon/src/audio/codecs/audiocodec.h
	daemon/src/audio/codecs/ulaw.cpp

+++ daemon/src/audio/codecs/ulaw.cpp
@@ -55,14 +55,13 @@ class Ulaw : public sfl::AudioCodec {
             return buf_size;
         }
 
-        SFLAudioSample ULawDecode(uint8 ulaw)
-        {
+        SFLAudioSample ULawDecode(uint8_t ulaw) {
             ulaw ^= 0xff;  // u-law has all bits inverted for transmission
             int linear = ulaw & 0x0f;
             linear <<= 3;
             linear |= 0x84;  // Set MSB (0x80) and a 'half' bit (0x04) to place PCM value in middle of range
 
-            uint shift = ulaw >> 4;
+            uint8_t shift = ulaw >> 4;
             shift &= 7;
             linear <<= shift;
             linear -= 0x84; // Subract uLaw bias
@@ -73,10 +72,9 @@ class Ulaw : public sfl::AudioCodec {
                 return linear;
         }
 
-        uint8 ULawEncode(SFLAudioSample pcm16)
-        {
+        uint8_t ULawEncode(SFLAudioSample pcm16) {
             int p = pcm16;
-            uint u;  // u-law value we are forming
+            uint8_t u;  // u-law value we are forming
 
             if (p < 0) {
                 p = ~p;
